% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%on\%}
\alias{\%on\%}
\title{Comparisons Operator}
\usage{
lhs \%on\% rhs
}
\arguments{
\item{lhs}{the left hand side of the operator, a vector}

\item{rhs}{the right hand side of the operator, a vector of numeric or string values}
}
\description{
This operator takes one variable and tests if it satisfies multiple conditions. 
It is related to \code{\%in\%} but has the additional feature of allowing for >, <, >=, and <= .
Use it just as you would \code{\%in\%}. It is named "on" because that is a common verb for
furniture (e.g., sit \emph{on} a chair, place it \emph{on} the table, etc.)
}
\examples{

x = c(1,2,3,4,3,2,1,8,9,7)
y = rnorm(10)
z = c("Yes", "No", "Yes", "No", "No", "Yes", "No", "No", "Yes", "No")
data = data.frame(x, y, z)

data$x \%on\% c(2, 4, ">5")
ifelse(data$x \%on\% c(2, 4, ">5"), 1, 0)

}

