% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%xt\%}
\alias{\%xt\%}
\title{Simple Crosstabs Operator}
\usage{
lhs \%xt\% rhs
}
\arguments{
\item{lhs}{the left hand side of the operator, a vector}

\item{rhs}{the right hand side of the operator, a vector}
}
\description{
This operator takes two variables and computes a simple cross tab.
}
\examples{

b = c(1,0,0,1,1,0,1,1,1,0)
x = c(1,2,3,2,3,3,1,0,0,0)
y = rnorm(10)
z = c("Yes", "No", "Yes", "No", "No", "Yes", "No", "No", "Yes", "No")
data = data.frame(b, x, y, z)

factor(data$x) \%xt\% factor(data$b)

}
