% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_wide.R
\name{long}
\alias{long}
\title{Wide to Long Data Reshaping}
\usage{
long(data, varying, ...)
}
\arguments{
\item{data}{the data.frame containing the wide format data}

\item{varying}{the variables that are time-varying that are to be placed in long format, 
needs to be in the format \code{list(c("x1", "x2"), c("z1", "z2"), etc.)}. If the data is 
unbalanced (e.g., there are three time points measured for one variable but only two for another),
using the placeholder variable \code{miss}, helps fix this.}

\item{...}{other arguments accepted by \code{reshape()}}
}
\description{
\code{long()} is a wrapper of \code{reshape()} that takes the data 
from a wide format to a long format. All arguments that \code{reshape()} accepts,
other than \code{direction} since it is set to "long" here, are accepted 
here as well.
}
\examples{

x1 <- runif(1000)
x2 <- runif(1000)
x3 <- runif(1000)
y1 <- rnorm(1000)
y2 <- rnorm(1000)
z  <- factor(sample(c(0,1), 1000, replace=TRUE))
a  <- factor(sample(c(1,2), 1000, replace=TRUE))
b  <- factor(sample(c(1,2,3,4), 1000, replace=TRUE))
df  <- data.frame(x1, x2, y1, y2, z, a, b)

## "Balanced" Data
ldf1 <- long(df, varying = list(c("x1", "x2"),
                                c("y1", "y2")),
             v.names = c("x", "y"))

## "Unbalanced" Data
ldf2 = long(df, 
            varying = list(c("x1", "x2", "x3"),
                           c("y1", "y2", "miss")),
            v.names = c("x", "y"))


}
\seealso{
\code{stats::reshape()} and \code{sjmisc::to_long()}
}
\author{
Tyson S. Barrett
}
