\name{print.frames}
\alias{print.frames}
\alias{print}
\title{
Print Method for the Frames Object
}
\description{
The print method for the frames object. By default, only the Average Marginal Effects are shown in the output from the frames object.
}
\usage{
print(x, ...)
}
\arguments{
  \item{x}{
the "frames" object
}
  \item{\dots}{
other arguments that can be applied to the print function
}
}
\value{
The data.frame of AME from the \code{frames()} function is printed.
}

\author{
Tyson S. Barrett
}

\seealso{
\code{\link{frames}}, \code{\link{tp2frames}}, \code{\link{summary.frames}}
}

\keyword{ print }% use one of  RShowDoc("KEYWORDS")
\keyword{ frames }% __ONLY ONE__ keyword per line
