\name{print.table1}
\alias{print.table1}
\title{Print Method for Table1}
\description{
The print method objects of class "table1" (from the \code{table1()} function). 
}
\usage{
print.table1(x, ...)
}
\arguments{
  \item{x}{
the "table1" object from the \code{table1()} function.
}
  \item{\dots}{
other arguments that can be applied to the print function
}
}

\value{
Prints the table1 object with the default row.names option switched to \code{row.names = FALSE}.
}

\author{
Tyson S. Barrett
}

\seealso{
\code{\link{table1}}
}

\keyword{ print }
\keyword{ table1 }
