% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowmeans.R
\name{rowsums}
\alias{rowsums}
\title{Get Row Sums}
\usage{
rowsums(..., na.rm = FALSE)
}
\arguments{
\item{...}{the variables to be included in the row sums}

\item{na.rm}{default is FALSE}
}
\value{
the row sums
}
\description{
Does what \code{rowSums()} does but without having to cbind the variables. Makes it easier to use
with the tidyverse
}
