% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1_utils.R
\name{selecting}
\alias{selecting}
\title{Selecting Function}
\usage{
selecting(d_, ...)
}
\arguments{
\item{d_}{the data.frame}

\item{...}{the variables}
}
\value{
The data.frame with the selected variables
}
\description{
For internal use in \code{table1()} and \code{tableC()} to extract the right data. 
Can also be used much like \code{dplyr::select()}, although I'd recommend
one to use \code{dplyr::select()} in general.
}
