\name{table1_}
\alias{table1_}
\title{Table 1 NSE}
\description{ Allows Non-standard evaluation of Table 1}

\usage{
table1_(d_, vars, index=index)
}

\arguments{
  \item{d_}{
      the data.frame that is to be summarized
}
  \item{vars}{
      formula or list of formulas containing the variables in the data set that are to be summarized
}
  \item{index}{
      to know whether the vars are an index
}
}

\author{Tyson Barrett and Emily Brignone}

\seealso{
\code{\link{print.table1}}
}

\keyword{ table1 }
\keyword{ descriptives }
\keyword{ summarize }
\keyword{ formatted }
