% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table1.R
\name{table1_format_condense}
\alias{table1_format_condense}
\title{Internal Table 1 Formatting Function (Condense)}
\usage{
table1_format_condense(d, tab, tab2, tests, test, NAkeep, rounding_perc,
  format_output, second, nams, simple, output, f1)
}
\arguments{
\item{d}{the data}

\item{tab}{the summary statistics}

\item{tab2}{the secondary summary statistics}

\item{tests}{the tests}

\item{test}{whether there should be significance tests performed}

\item{NAkeep}{logical; should we keep the NA's in factor variables}

\item{rounding_perc}{the number of decimal places to round the percentages}

\item{format_output}{the output to be displayed (see \code{table1()})}

\item{second}{the list of variables to apply FUN2}

\item{nams}{the variables to which FUN2 is to be applied}

\item{simple}{only percentages to be produced?}

\item{output}{how to print the table (see \code{table1()})}

\item{f1}{the formatting of the numbers}
}
\value{
A data.frame
}
\description{
For internal use in table1() to format table with condensing.
}
