% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table1.R
\name{table1_summarizing}
\alias{table1_summarizing}
\title{Internal Table 1 Summarizing Function}
\usage{
table1_summarizing(d, num_fun, num_fun2, second, row_wise, test, NAkeep)
}
\arguments{
\item{d}{the data}

\item{second}{the variables to which FUN2 is to be applied}

\item{row_wise}{}

\item{test}{}

\item{FUN}{the summarizing function}

\item{FUN2}{the summarizing function}
}
\value{
A data.frame
}
\description{
For internal use in table1().
}
