% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableF.R
\name{tableF}
\alias{tableF}
\title{Frequency Table}
\usage{
tableF(.data, x, n = 20, splitby = NULL)
}
\arguments{
\item{.data}{the data frame containing the variable}

\item{x}{the bare variable name (not quoted)}

\item{n}{the number of values shown int he table}

\item{splitby}{the stratifying variable}
}
\value{
a list of class \code{tableF} containing the frequency table(s)
}
\description{
Provides in-depth frequency counts and percentages.
}
