% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_latex2.R
\name{to_latex}
\alias{to_latex}
\title{From Table 1 to Latex 2}
\usage{
to_latex(tab, caption, align, len, splitby, float, booktabs, cor_type = NULL)
}
\arguments{
\item{tab}{the table1 object}

\item{caption}{caption character vector}

\item{align}{align character vector}

\item{len}{the number of levels of the grouping factor}

\item{splitby}{the name of the grouping factor}

\item{float}{argument for latex formatting}

\item{booktabs}{add booktabs to latex table?}

\item{cor_type}{optional argument regarding the correlation type (for tableC)}
}
\description{
Internal \code{table1()} and \code{tableC()} function for providing output = "latex2"
}
