% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_wide.R
\name{wide}
\alias{wide}
\title{Long to Wide Data Reshaping}
\usage{
wide(data, v.names, timevar, idvar = NULL, ...)
}
\arguments{
\item{data}{the data.frame containing the wide format data}

\item{v.names}{the variable names in quotes of the measures to be separated into multiple columns based on the time variable}

\item{timevar}{the variable name in quotes of the time variable}

\item{idvar}{the ID variable name in quotes}

\item{...}{other arguments accepted by \code{reshape()}}
}
\description{
\code{wide()} is a wrapper of \code{reshape()} that takes the data 
from a long format to a wide format. All arguments that \code{reshape()} accepts,
other than \code{direction} since it is set to "wide" here, are accepted 
here as well.
}
\seealso{
\code{stats::reshape()}
}
\author{
Tyson S. Barrett
}
