% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_wide.R
\name{long}
\alias{long}
\title{Wide to Long Data Reshaping}
\usage{
long(data, varying, ...)
}
\arguments{
\item{data}{the data.frame containing the wide format data}

\item{varying}{the variables that are time-varying that are to be placed in long format, 
needs to be in the format \code{list(c("x1", "x2"), c("z1", "z2"), etc.)}}

\item{...}{other arguments accepted by \code{reshape()}}
}
\description{
\code{long()} is a wrapper of \code{reshape()} that takes the data 
from a wide format to a long format. All arguments that \code{reshape()} accepts,
other than \code{direction} since it is set to "long" here, are accepted 
here as well.
}
\seealso{
\code{stats::reshape()} and \code{sjmisc::to_long()}
}
\author{
Tyson S. Barrett
}
