% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowmeans.R
\name{rowmeans}
\alias{rowmeans}
\title{Get Row Means}
\usage{
rowmeans(..., na.rm = FALSE)
}
\arguments{
\item{...}{the variables to be included in the row means}

\item{na.rm}{default is FALSE}
}
\value{
the row means
}
\description{
Does what \code{rowMeans()} does but without having to cbind the variables. Makes it easier to use
with the tidyverse
}
