% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowmeans.R
\name{rowsums}
\alias{rowsums}
\title{Get Row Sums}
\usage{
rowsums(..., na.rm = FALSE)
}
\arguments{
\item{...}{the values in the variable that are to be replaced by either NA's or the value set by the user. Can be a function (or multiple functions) to specify values to change (e.g., is.nan(), is.na()).}

\item{na.rm}{default is FALSE}
}
\value{
the row means
}
\description{
Does what \code{rowSums()} does but without having to cbind the variables. Makes it easier to use
with the tidyverse
}
