% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table1.R
\name{table1_}
\alias{table1_}
\title{Internal Table 1 Function}
\usage{
table1_(..., d_, .cl = NULL)
}
\arguments{
\item{...}{the variables}

\item{d_}{the data.frame}

\item{.cl}{the original functon call}
}
\value{
A data.frame
}
\description{
For internal use in table1() to extract the right data.
}
