% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table1.R
\name{table1_}
\alias{table1_}
\title{Table 1 for Health, Behavioral, and Social Scientists}
\usage{
table1_(d_, vars, split = FALSE)
}
\arguments{
\item{d_}{the data.frame}

\item{vars}{uses dots_capture() to capture the ...}

\item{split}{for internal use in table1()}
}
\value{
A data.frame
}
\description{
Produces a descriptive table, stratified by an optional categorical variable, 
providing means/frequencies and standard deviations/percentages. 
It is well-formatted for easy transition to academic article or report. 
Can be used within the piping framework [see library(magrittr)].
}

