% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table1.R
\name{table1_summarizing}
\alias{table1_summarizing}
\title{Internal Table 1 Summarizing Function}
\usage{
table1_summarizing(d, num_fun, num_fun2, second, row_wise, test, NAkeep)
}
\arguments{
\item{d}{the data}

\item{num_fun}{the summarizing function}

\item{num_fun2}{the second summarizing function}

\item{second}{the variables to which FUN2 is to be applied}

\item{row_wise}{the way to compute the percentages}

\item{test}{should significance tests be run?}

\item{NAkeep}{whether NA's should be shown in the output of categorical variables}
}
\value{
A data.frame
}
\description{
For internal use in table1() to summarize data.
}
